routeId overrides from id
As routeId is set only after the from id, it makes unconsistent the
default ids that are using the routeId. Instead of reprocessing the
whole route, I prefererd to just override the from id once the routeId
is defined. Of course, this applies only if you have defined some
NamingStrategy.